+++ srcs/create_empty_stack.c
@@ -6,7 +6,7 @@
 /*   By: jleroux <jleroux@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/28 16:40:20 by jleroux           #+#    #+#             */
-/*   Updated: 2022/06/29 13:45:22 by jleroux          ###   ########.fr       */
+/*   Updated: 2022/06/29 15:02:38 by jleroux          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ t_stack	create_empty_stack(int size)
 {
 	t_stack	stack;
 
-	stack.arr = (long int *)malloc(sizeof(long int) * size);
+	stack.arr = (long long *)malloc(sizeof(long long) * size);
 	stack.size = 0;
 	return (stack);
 }
